+++ packages/apps/human-app/frontend/src/main.tsx
@@ -19,7 +19,7 @@ import { JWTExpirationCheck } from '@/shared/contexts/jwt-expiration-check';
 import { ColorModeProvider } from '@/shared/contexts/color-mode';
 import { HomePageStateProvider } from '@/shared/contexts/homepage-state';
 import { RegisteredOraclesProvider } from '@/shared/contexts/registered-oracles';
-import { NotificationProvider } from '@/shared/providers/notifications-provider';
+import { NotificationProvider } from './shared/contexts/notifications-provider';
 
 const root = document.getElementById('root');
 if (!root) throw Error('root element is undefined');
