+++ packages/apps/human-app/frontend/src/modules/homepage/components/welcome.tsx
@@ -12,12 +12,12 @@ import { useIsMobile } from '@/shared/hooks/use-is-mobile';
 import { OperatorSignIn } from '@/modules/homepage/hooks/use-operator-signin';
 import { WorkerSignIn } from '@/modules/homepage/components/worker-signin';
 import { useColorMode } from '@/shared/contexts/color-mode';
-import { useBackgroundColorStore } from '@/shared/hooks/use-background-store';
 import { useHomePageState } from '@/shared/contexts/homepage-state';
+import { useBackgroundColorContext } from '@/shared/hooks/use-background-color-context';
 
 export function Welcome() {
   const { colorPalette, isDarkMode } = useColorMode();
-  const { setWhiteBackground } = useBackgroundColorStore();
+  const { setWhiteBackground } = useBackgroundColorContext();
   const { setPageView } = useHomePageState();
   const { t } = useTranslation();
   const logoText: string = t('homepage.humanApp');
