+++ packages/apps/human-app/frontend/src/shared/components/layout/unprotected/layout.tsx
@@ -1,6 +1,6 @@
 import { Container, Grid } from '@mui/material';
 import { Outlet } from 'react-router-dom';
-import { useBackgroundColorStore } from '@/shared/hooks/use-background-store';
+import { useBackgroundColorContext } from '@/shared/hooks/use-background-color-context';
 import { useIsMobile } from '@/shared/hooks/use-is-mobile';
 import { breakpoints } from '@/shared/styles/breakpoints';
 import { useColorMode } from '@/shared/contexts/color-mode';
@@ -13,7 +13,7 @@ interface LayoutProps {
 
 export function Layout({ withNavigation = true }: LayoutProps) {
   const { colorPalette, isDarkMode } = useColorMode();
-  const { backgroundColor } = useBackgroundColorStore();
+  const { backgroundColor } = useBackgroundColorContext();
   const isMobile = useIsMobile();
   const layoutBackgroundColor = (() => {
     if (isDarkMode) {
