+++ packages/apps/human-app/frontend/src/shared/contexts/color-mode/color-mode.tsx
@@ -11,7 +11,7 @@ import { theme } from '@/shared/styles/theme';
 import { colorPalette as defaultColorPalette } from '@/shared/styles/color-palette';
 import { darkTheme } from '@/shared/styles/dark-theme';
 import { darkColorPalette } from '@/shared/styles/dark-color-palette';
-import { BackgroundProvider } from '@/shared/contexts/background-color-store';
+import { BackgroundProvider } from '../background';
 import {
   ColorMode,
   hasColorMode,
